+++ src/components/LandingPageMapTab.tsx
@@ -7,11 +7,6 @@ import Divider from "@mui/material/Divider";
 import LandingPageMap from "./LandingPageMap";
 import AllProgramMap from "./AllProgramMap";
 import allPrograms from "../data/allPrograms.json";
-import commodities from "../images/legends/Commodities benefits 2018 - 2022.png";
-import conservation from "../images/legends/conservation programs benefits 2018 - 2022.png";
-import crop from "../images/legends/crop insurance 2018 - 2022.png";
-import snap from "../images/legends/SNAP programs benefits 2018 - 2022.png";
-import total from "../images/legends/total farm bill benefits 2018 - 2022.png";
 import LandingDisplay from "./LandingDisplay";
 
 const theme = createTheme({
@@ -30,25 +25,6 @@ interface TabPanelProps {
 
 function TabPanel(props: TabPanelProps) {
 	const { value, index, title, ...other } = props;
-	let colorLegend = null;
-
-	switch (title) {
-	case "Title I: Commodities":
-		colorLegend = commodities;
-		break;
-	case "Title II: Conservation":
-		colorLegend = conservation;
-		break;
-	case "Crop Insurance":
-		colorLegend = crop;
-		break;
-	case "Supplemental Nutrition Assistance Program (SNAP)":
-		colorLegend = snap;
-		break;
-	case "All Programs":
-		colorLegend = total;
-		break;
-	}
 
 	return (
 		<Box role="tabpanel" hidden={value !== index} {...other}>
@@ -59,9 +35,6 @@ function TabPanel(props: TabPanelProps) {
 						mx: "auto"
 					}}
 				>
-					<Box sx={{ display: "flex", justifyContent: "center", mt: 2 }}>
-						<CardMedia sx={{ maxWidth: 720, mt: 3 }} component="img" src={colorLegend} />
-					</Box>
 					{title === "All Programs" ? <AllProgramMap /> : <LandingPageMap programTitle={title} />}
 				</Box>
 			)}
@@ -147,7 +120,7 @@ export default function LandingPageMapTab(): JSX.Element {
     							</Box>
     						}
     					/>
-    					<Divider sx={{ mx: 1 }} orientation="vertical" variant="middle" flexItem />
+    					{/* <Divider sx={{ mx: 1 }} orientation="vertical" variant="middle" flexItem />
     					<CustomTab
     						label={
     							<Box>
@@ -156,7 +129,7 @@ export default function LandingPageMapTab(): JSX.Element {
     								<Typography>${Number(cropTotal / 1000000000.0).toFixed(2)}B</Typography>
     							</Box>
     						}
-    					/>
+    					/> */}
     					<Divider sx={{ mx: 1 }} orientation="vertical" variant="middle" flexItem />
     					<CustomTab
     						label={
@@ -174,8 +147,8 @@ export default function LandingPageMapTab(): JSX.Element {
     		<TabPanel value={value} index={2} title="All Programs" />
     		<TabPanel value={value} index={4} title="Title I: Commodities" />
     		<TabPanel value={value} index={6} title="Title II: Conservation" />
-    		<TabPanel value={value} index={8} title="Crop Insurance" />
-    		<TabPanel value={value} index={10} title="Supplemental Nutrition Assistance Program (SNAP)" />
+    		{/* <TabPanel value={value} index={8} title="Crop Insurance" />*/}
+    		<TabPanel value={value} index={8} title="Supplemental Nutrition Assistance Program (SNAP)" />
     	</Box>
     );
 }